Fix gcc-4 compilation of addrfilt.c
John Hasler sent in a patch to do this (which still wouldn't make it compile
for me).  This reminded me that I had tackled this myself when my distro moved
to gcc-4 a while back.  It turned out I had never even checked in the file from
the working copy I was using (!).  Anyway, here it is now.